+++ packages/expo-audio/android/src/main/java/expo/modules/audio/AudioRecorder.kt
@@ -18,7 +18,7 @@ import expo.modules.kotlin.sharedobjects.SharedObject
 import java.io.File
 import java.io.IOException
 import java.util.UUID
-import kotlin.math.ln
+import kotlin.math.log10
 
 private const val RECORDING_STATUS_UPDATE = "recordingStatusUpdate"
 
@@ -40,16 +40,16 @@ class AudioRecorder(
   var isRecording = false
   var isPaused = false
 
-  private fun getAudioRecorderLevels(): Int? {
+  private fun getAudioRecorderLevels(): Double? {
     if (!meteringEnabled || recorder == null || !isRecording) {
       return null
     }
 
     val amplitude: Int = recorder?.maxAmplitude ?: 0
     return if (amplitude == 0) {
-      -160
+      -160.0
     } else {
-      (20 * ln(amplitude.toDouble() / 32767.0)).toInt()
+      20 * log10(amplitude.toDouble() / 32767.0)
     }
   }
 
@@ -168,7 +168,7 @@ class AudioRecorder(
       putBoolean("isRecording", isRecording)
       putLong("durationMillis", getAudioRecorderDurationMillis())
       getAudioRecorderLevels()?.let {
-        putInt("metering", it)
+        putDouble("metering", it)
       }
       filePath?.let {
         val path = Uri.fromFile(File(it)).toString()