+++ packages/expo-system-ui/CHANGELOG.md
@@ -14,6 +14,10 @@
 
 - Added support for React Native 0.80.x. ([#37400](https://github.com/expo/expo/pull/37400) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 
+## 5.0.9 - 2025-06-18
+
+_This version does not introduce any user-facing changes._
+
 ## 5.0.8 - 2025-06-04
 
 _This version does not introduce any user-facing changes._